Right-size the backend test suite
The per-batch pass over the whole suite (real Redis landed as #63):
33 tests and 599 net lines gone across sandbox, mcp, notifications,
harnesses, api+identity, extensions+doctor, usage+settings, and
platform misc — mock theater, change-detectors, trivial-code tests,
permutation padding, duplicate-altitude coverage, zero-consumer
fixtures, and mode→extension rename artifacts. build+scope, durable
core, and webhooks needed nothing.

Rides along, from chasing this PR's CI:
- the mcp disconnect test read the cache through a portal-bound client
  from the test's own loop — cross-loop reads flake or park forever;
  it now abandons the shared client and dials fresh (the same latent
  bug from #63 was failing other branches' CI too)
- a suite watchdog in conftest: nine minutes into a ~90s suite,
  faulthandler dumps every thread's stack and exits from inside the
  process, so the pre-existing durable-module hang self-reports with
  frames instead of stalling a runner for hours; the job also gets a
  15-minute outer timeout
